generate a year folder if absent in the archive
In previous commits, each year of the archive had to be defined
once by the user, or move2archive did not process the file to
archive.  By the present commit, move2archive attempts the set up
the yearly archive on the fly.  In the same run, the file to store
can be moved into this directory, too.